% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Concordance.R
\docType{methods}
\name{audit}
\alias{audit}
\alias{audit.RedcapDqa}
\title{Concordance Checks}
\usage{
audit(obj, ...)

\S4method{audit}{RedcapDqa}(obj, ...)
}
\arguments{
\item{obj}{A \code{RedcapDqa} object}
}
\value{
A \code{RedcapDqaResult}.

This contains:

\describe{
\item{concordance}{Overall concordance}
\item{record.concordance}{Concordance for each record}
\item{field.concordance}{Concordance for each variable}
\item{matrix.concordance}{Boolean matrix of concordance for each data cell}
}
}
\description{
Run crude concordance checks
}
\details{
This utility function instantiates the DQA process.

It pulls the record ids from both repositories and matches the common identifiers.

Data is then streamed from both repos and then matched for common fields.

These datasets are used to create an instance of a \code{RedcapDqa} object.
}
\seealso{
\code{\link{RedcapDqa}}
}
\concept{
RedcapDqa
}

