% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper.R
\name{copy_from_env}
\alias{copy_from_env}
\title{Copy environments}
\usage{
copy_from_env(from, to)
}
\arguments{
\item{from}{Environment to copy from.}

\item{to}{Environment to append to.}

\item{objs}{Objects to copy. If not specified, this generalizes to all objects in the from environment}
}
\description{
Copy all entries in one environment and append to another.
}
\details{
This function copies objects from one environment to another.

This helps in transfer of data as functions may be nested inside each other but may need to manipulate each other's data.

This is not in line with FP due to side effects and mutations that can arise but we have implemented this for computational convinience and separation of concerns
}
\concept{RedcapDqa}
