% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Concordance.R
\name{document.RedcapDqaResult}
\alias{document.RedcapDqaResult}
\alias{document.RedcapDqaResults}
\title{Export DQA Results to Word}
\usage{
document.RedcapDqaResults(x, report.location = tempfile(pattern = "dqa",
  tmpdir = tempdir(), fileext = ".docx"), ...)
}
\description{
Export DQA Results to Word
}
\details{
Export DQA Results to Word for easier interpretation and visualization.
}
\concept{
RedcapDqa
}

