% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RedcapDqa.R
\docType{methods}
\name{print}
\alias{print}
\alias{print.RedcapDqa}
\title{Print DQA object}
\usage{
print(x, ...)

\S4method{print}{RedcapDqa}(x, ...)
}
\description{
Print DQA object
}
\details{
Generic Print method
}
\concept{
RedcapDqa
}

