% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Concordance.R
\name{print.RedcapDqaResult}
\alias{print.RedcapDqaResult}
\title{print DQA Results}
\usage{
print.RedcapDqaResult(x, ...)
}
\description{
print DQA Results
}
\details{
Generic print method
}
\concept{RedcapDqa}
