% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Concordance.R
\name{print.RedcapDqaResults}
\alias{print.RedcapDqaResults}
\title{print DQA Results}
\usage{
print.RedcapDqaResults(x, ...)
}
\description{
print DQA Results
}
\details{
Generic print method
}
\concept{
RedcapDqa
}

