% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{RedcapDqa}
\alias{RedcapDqa}
\title{Redcap Data Quality Assurance}
\description{
Class to hold DQA datasets and other allied objects
}
\details{
This class holds the DQA and repository datasets as well other objects that are created in the process of DQA.

It stores these allied objects in its own cache and has a log to track events that occur in the process of DQA
}
\section{Slots}{

\describe{
\item{\code{dqaData}}{DQA dataset}

\item{\code{repoData}}{Repository dataset}

\item{\code{cache}}{Temporary storage}
}}
\concept{
RedcapDqa
}

