% Generated by roxygen2 (4.0.2): do not edit by hand
\name{redcap_dqa}
\alias{redcap_dqa}
\title{Instantiate DQA}
\usage{
redcap_dqa(repo_token, dqa_token,
  repo_api_url = "http://localhost/redcap/api/", dqa_api_url = repo_api_url,
  local = T, dqa_local = local, repo_local = local)
}
\arguments{
\item{repo_token}{Data repository's token}

\item{dqa_token}{DQA repository's token}

\item{repo_api_url}{Url to data repo's api}

\item{dqa_api_url}{Url to DQA repo's api}

\item{local}{Flag as to whether the repo's are local to the R instance}

\item{dqa_local}{Flag as to whether the dqa repo is local to the R instance}

\item{repo_local}{Flag as to whether the data repo is local to the R instance}
}
\value{
A \code{RedcapDqa} object loaded with data.
}
\description{
Wrapper function for instantiating DQA objects
}
\details{
This utility function instantiates the DQA process.

It pulls the record ids from both repositories and matches the common identifiers.

Data is then streamed from both repos and then matched for common fields.

These datasets are used to create an instance of a \code{RedcapDqa} object.
}
\seealso{
\code{\link{RedcapDqa}}
}
\concept{
RedcapDqa
}

