% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Concordance.R
\name{show.RedcapDqaResults}
\alias{show.RedcapDqaResults}
\title{show DQA Results}
\usage{
show.RedcapDqaResults(x, ...)
}
\description{
show DQA Results
}
\details{
Generic show method
}
\concept{RedcapDqa}
