% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RedcapDqa.R
\docType{methods}
\name{show.RedcapDqa}
\alias{show.RedcapDqa}
\title{Show DQA object}
\usage{
\S4method{show}{RedcapDqa}(object)
}
\description{
Show DQA object
}
\details{
Generic show method
}
\concept{RedcapDqa}
