% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Concordance.R
\name{summary.RedcapDqaResult}
\alias{summary.RedcapDqaResult}
\title{summary DQA Results}
\usage{
\method{summary}{RedcapDqaResult}(x, ...)
}
\description{
Summarize DQA Results
}
\details{
Generic summary method
}
\concept{
RedcapDqa
}

