% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper.R
\name{validate_inputs}
\alias{validate_input}
\alias{validate_inputs}
\title{Validate inputs}
\usage{
validate_input(env)
}
\arguments{
\item{cxt}{Context representionf the current context of the function execution. Can be a list or an environment.}
}
\description{
Check the inputs that control the dqa process.
}
\details{
This utility function checks the input fed into the wrapper function that instantiates the RedcapDqa object.

It checks the validity of the input and flags any errors encountered.

This is to separate the concerns of the wrapper function and make it leaner
}
\seealso{
\code{\link{redcap_dqa}}
}
\concept{
RedcapDqa
}

