% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper.R
\name{wrangle_audit_data}
\alias{wrangle_audit_data}
\title{Wrangle audit data}
\usage{
wrangle_audit_data(env)
}
\arguments{
\item{env}{Environment of the calling function.}
}
\description{
Wrangle the data to be used for the audit process.
}
\details{
This utility function wrangles the data to be used in the validation process.

This includes subsetting to only include fields included in both datasets in addition to the inclusion and exclusion of specific fields as specified in the wrapper function

This also separates the concerns of the main wrapper function in addition to making it leaner
}
\seealso{
\code{\link{redcap_dqa}}
}
\concept{
RedcapDqa
}

