% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wrappers.R
\name{redcap_dqa}
\alias{redcap_dqa}
\title{Instantiate DQA object}
\usage{
redcap_dqa(repo_file_location = NA, dqa_file_location = NA,
  meta_file_location = NA, repo_token = NA, dqa_token = NA,
  repo_api_url = "http://localhost/redcap/api/",
  dqa_api_url = repo_api_url, local = T, dqa_local = local,
  repo_local = local, fields_to_include = NA, fields_to_exclude = NA,
  id_var = NA, min_date = NA, max_date = NA,
  date_var = "date_today", strata = NA, site_id = NA)
}
\arguments{
\item{repo_file_location}{Repository data file location. If specified, data is not pulled from the repository.}

\item{dqa_file_location}{DQA data file location. If specified, data is not pulled from the repository.}

\item{repo_token}{Data repository's token.}

\item{dqa_token}{DQA repository's token.}

\item{repo_api_url}{Url to data repo's api.}

\item{dqa_api_url}{Url to DQA repo's api.}

\item{local}{Flag as to whether the repo's are local to the R instance.}

\item{dqa_local}{Flag as to whether the dqa repo is local to the R instance.}

\item{repo_local}{Flag as to whether the data repo is local to the R instance.}

\item{fields_to_include}{Fields to be includes in the audit process. If included, then the fields to be excluded will be ignored.}

\item{fields_to_exclude}{Fields to be excluded from the audit process. For example, timestamps. If fields to be included are specified, then this variable is ignored}

\item{id_var}{Field to be used to identify records across the two repositories.}

\item{min_date}{Minimum date used to constrain data.}

\item{max_date}{Maximum date used to constrain data.}
}
\value{
A \code{RedcapDqa} object loaded with data.
}
\description{
Wrapper function for instantiating DQA objects
}
\details{
This utility function instantiates the DQA process.

Data is either pulled from the repositories or read in from files.

The datasets are then matched for common fields and identifiers.

The data and other metadata relevant to the audit process are then stored into an object for consequent analysis.

This results to an instance of a \code{RedcapDqa} object.
}
\seealso{
\code{\link{RedcapDqa}}
}
\concept{RedcapDqa}
